% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_corpus.R
\name{create_corpus}
\alias{create_corpus}
\title{Create a corpus}
\usage{
create_corpus(path, encoding = "UTF-8")
}
\arguments{
\item{path}{A string containing the path to a folder of plain text files (ending in .txt) with their name structured as following: authorname_textname.txt (e.g. smith_text1.txt).}

\item{encoding}{Either a single string indicating the encoding for all files or a vector of strings indicating the encodings for each file. The default is UTF-8.}
}
\value{
A \code{quanteda} corpus object with the authors' names and the text names as docvars.
}
\description{
Function to read in text data and turn it into a \code{quanteda} corpus object.
}
\examples{
\dontrun{
path <- "path/to/data"
create_corpus(path)
}

}
