% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enron.sample.R
\docType{data}
\name{enron.sample}
\alias{enron.sample}
\title{Enron sample}
\format{
A \code{quanteda} corpus object.
}
\source{
Halvani, Oren. 2021. Practice-Oriented Authorship Verification. Technical University of Darmstadt PhD Thesis.
https://tuprints.ulb.tu-darmstadt.de/19861/
}
\usage{
enron.sample
}
\description{
A small sample of the \emph{Enron} corpus comprising ten authors with approximately the same amount of data. Each author has one text labelled as 'unknown' and the other texts labelled as 'known'. The data was pre-processed using the \emph{POSnoise} algorithm to mask content (see \code{\link[=contentmask]{contentmask()}}).
}
\keyword{datasets}
