% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/important_features.R
\name{important_features}
\alias{important_features}
\title{Important features}
\usage{
important_features(q, candidate, impostors)
}
\arguments{
\item{q}{This is one \code{quanteda} dfm row corresponding to the Q text.}

\item{candidate}{This is the \code{quanteda} dfm corresponding to the candidate text(s).}

\item{impostors}{This is the \code{quanteda} dfm with the impostors' data.}
}
\value{
One named vector of features that belong to the overlap between the Q text and the candidate containing the ratio between candidate and impostors
}
\description{
This function shows the most important features used by any of the impostors algorithms that adopt the minmax distance.
}
\examples{
q <- "aaabbc"
candidate <- "aaaaaaaabbbbccc"
impostor <- "abbccccccccccczzzz"
corpus <- quanteda::corpus(c(q, candidate, impostor))
d <- vectorize(corpus, n = 1)
important_features(d[1,], d[2,], d[3,])
}
