% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impostors.R
\name{impostors}
\alias{impostors}
\title{Impostors Method}
\usage{
impostors(
  qs,
  candidates,
  cand.imps,
  q.imps,
  algorithm = "RBI",
  coefficient = "minmax",
  k = 300,
  m = 100,
  n = 25,
  cores = NULL
)
}
\arguments{
\item{qs}{The \code{quanteda} dfm containing the disputed texts to test. With the exception of the "IM" algorithm, this data frame can contain the same texts that are also in the candidate data frame, for example, for leave-one-out testing.}

\item{candidates}{The \code{quanteda} dfm containing the data belonging to the candidate authors to test.}

\item{cand.imps}{The \code{quanteda} dfm containing the impostors (or only the impostors for the candidate data if the algorithm is IM)}

\item{q.imps}{The \code{quanteda} dfm containing the impostors for the disputed text (only applicable for the IM algorithm)}

\item{algorithm}{A string specifying which impostors algorithm to use, either "RBI", "KGI", or "IM".}

\item{coefficient}{A string indicating the coefficient to use, either "minmax" (default) or "cosine". This does not apply to the algorithm KGI, where the distance is "minmax".}

\item{k}{The \emph{k} parameters for the RBI algorithm. Not used by other algorithms. The default is 300.}

\item{m}{The \emph{m} parameter for the IM algorithm. Not used by other algorithms. The default is 100.}

\item{n}{The \emph{n} parameter for the IM algorithm. Not used by other algorithms. The default is 25.}

\item{cores}{The number of cores to use for parallel processing (the default is one).}
}
\value{
The function will test all possible combinations of q texts and candidate authors and return a
data frame containing the score ranging from 0 to 1 representing the degree of confidence that the candidate is the author of the Q text. The data frame contains a column called "target" with a logical value which is TRUE if the author of the Q text is the candidate and FALSE otherwise. If the RBI algorithm is selected then the data frame will also contain a column with the features that are likely to have had an impact on the score.
}
\description{
This function runs the authorship verification method \emph{Impostors Method}.
}
\details{
More details here.
}
\examples{
library(quanteda)
text1 <- "The cat sat on the mat."
text2 <- "The dog sat on the mat."
k.corpus <-  corpus(c(text1, text2))
k.dfm <- dfm(tokens(k.corpus)) |> dfm_weight(scheme = "prop")
docvars(k.dfm, "author") <- "K"

imp1 <- "A pet, such as a cat or a dog, is an animal kept primarily for a person's company."
imp2 <- "Some pets may be accepted by the owner regardless of these characteristics."
imp.corpus <- corpus(c(imp1, imp2))
imp.dfm <- dfm(tokens(imp.corpus)) |> dfm_weight(scheme = "prop")
docvars(imp.dfm, "author") <- "wikipedia"

q <- "The dog sat on the chair."
q.dfm <- dfm(tokens(q)) |> dfm_weight(scheme = "prop")
docvars(q.dfm, "author") <- "Q"

#note how all the dfms must have the same columns
imp.dfm <- dfm_match(imp.dfm, featnames(k.dfm))
q.dfm <- dfm_match(q.dfm, featnames(k.dfm))

results <- impostors(q.dfm, k.dfm, imp.dfm, algorithm = "KGI")
}
