% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambdaG_visualize.R
\name{lambdaG_visualize}
\alias{lambdaG_visualize}
\title{Visualize the output of the LambdaG algorithm}
\usage{
lambdaG_visualize(
  q.data,
  k.data,
  ref.data,
  N = 10,
  r = 30,
  output = "html",
  print = TRUE,
  cores = NULL
)
}
\arguments{
\item{q.data}{A single questioned or disputed text as a corpus object (the output of \code{\link[=create_corpus]{create_corpus()}}).}

\item{k.data}{A known or undisputed corpus containing exclusively one candidate author's texts (the output of \code{\link[=create_corpus]{create_corpus()}}).}

\item{ref.data}{The reference dataset as a corpus (the output of \code{\link[=create_corpus]{create_corpus()}}).}

\item{N}{The order of the model. Default is 10.}

\item{r}{The number of iterations. Default is 30.}

\item{output}{A string detailing the file type of the colour-coded text output. Either "html" (default) or "latex".}

\item{print}{A logical value indicating whether the colour-coded text file should be written to the working directory (default) or not.}

\item{cores}{The number of cores to use for parallel processing (the default is one).}
}
\value{
The function outputs a list of two objects: a data frame with each row being a token in the Q text and the values of lambdaG for the token and sentences, in decreasing order of sentence lambdaG; the raw code in html or latex that generates the colour-coded file. If the print value is set to TRUE the function will also save the colour-coded text as an html or plain text file on disk in the working directory.
}
\description{
This function outputs a colour-coded list of sentences belonging to the input Q text ordered from highest to lowest lambdaG value, as shown in Nini et al. (pending submission).
}
\examples{
q.data <- corpus_trim(enron.sample[1], "sentences", max_ntoken = 10)
k.data <- enron.sample[2:10]
ref.data <- enron.sample[11:ndoc(enron.sample)]
outputs <- lambdaG_visualize(q.data, k.data, ref.data, r = 2, print = FALSE)
outputs$table
}
\references{
Nini, A., Halvani, O., Graner, L., Gherardi, V., Ishihara, S. Authorship Verification based on the Likelihood Ratio of Grammar Models. https://arxiv.org/abs/2403.08462v1
}
