% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refcor.sample.R
\docType{data}
\name{refcor.sample}
\alias{refcor.sample}
\title{Refcor sample}
\format{
\subsection{\code{who}}{

A \code{quanteda} corpus object created using \code{\link[=create_corpus]{create_corpus()}}.
}
}
\source{
\url{https://github.com/cophi-wue/refcor}
}
\usage{
refcor.sample
}
\description{
A small sample of the English section of the \emph{refcor} corpus made up of the first 1,000-1,500 words of the first chapter of three novels for three authors.
}
\keyword{datasets}
