% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk_texts.R
\name{chunk_texts}
\alias{chunk_texts}
\title{Chunk a corpus}
\usage{
chunk_texts(corpus, size)
}
\arguments{
\item{corpus}{A \code{quanteda} corpus.}

\item{size}{The size of the chunks in number of tokens.}
}
\value{
A \code{quanteda} corpus object where each text is a chunk of the size requested.
}
\description{
This function can be used to chunk a corpus in order to control sample sizes.
}
\examples{
corpus <- quanteda::corpus(c("The cat sat on the mat", "The dog sat on the chair"))
quanteda::docvars(corpus, "author") <- c("A", "B")
chunk_texts(corpus, size = 2)


}
