% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contentmask.R
\name{contentmask}
\alias{contentmask}
\title{Content masking}
\usage{
contentmask(parsed.corpus, algorithm = "POSnoise")
}
\arguments{
\item{parsed.corpus}{A data frame containing a \code{spacyr} Part of Speech tagged corpus. Typically the output of the \code{\link[=create_corpus]{create_corpus()}} function with the pos_tag argument switched to TRUE.}

\item{algorithm}{A string, either "POSnoise" (default) or "frames".}
}
\value{
A \code{quanteda} corpus object only containing functional tokens, depending on the algorithm chosen. The corpus also contains the author names as a docvar.
}
\description{
This function offers two algorithms for topic/content masking. See Details below.
}
\details{
References to algorithms to appear here.
}
\examples{
text <- "The elegant cat was forcefully put on the chair."
spacyr::spacy_initialize()
parsed.text <- spacyr::spacy_parse(text)
spacyr::spacy_finalize()
posnoise.text <- contentmask(parsed.text, algorithm = "POSnoise")
frames.text <- contentmask(parsed.text, algorithm = "framenoise")
}
