% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contentmask.R
\name{contentmask}
\alias{contentmask}
\title{Content masking}
\usage{
contentmask(corpus, algorithm = "POSnoise")
}
\arguments{
\item{corpus}{A \code{quanteda} corpus object, typically the output of the \code{\link[=create_corpus]{create_corpus()}} function.}

\item{algorithm}{A string, either "POSnoise" (default) or "frames".}
}
\value{
A \code{quanteda} corpus object only containing functional tokens, depending on the algorithm chosen. The corpus contains the same docvars as the input.
}
\description{
This function offers two algorithms for topic/content masking. See Details below.
}
\details{
References to algorithms to appear here.
}
\examples{
text <- "The elegant cat was forcefully put on the chair."
toy.corpus <- quanteda::corpus(text)
contentmask(toy.corpus, algorithm = "POSnoise")
contentmask(toy.corpus, algorithm = "framenoise")
}
