% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contentmask.R
\name{contentmask}
\alias{contentmask}
\title{Content masking}
\usage{
contentmask(
  corpus,
  model = "en_core_web_sm",
  algorithm = "POSnoise",
  replace_non_ascii = F
)
}
\arguments{
\item{corpus}{A \code{quanteda} corpus object, typically the output of the \code{\link[=create_corpus]{create_corpus()}} function.}

\item{model}{The spacy model to use. The default is en_core_web_sm.}

\item{algorithm}{A string, either "POSnoise" (default) or "frames".}

\item{replace_non_ascii}{A boolean value. If this is TRUE then All non-ASCII characters are either substituted to ASCII ones or removed using the function \code{\link[textclean:replace_non_ascii]{textclean::replace_non_ascii()}}. This operation also removes all emojis.}
}
\value{
A \code{quanteda} corpus object only containing functional tokens, depending on the algorithm chosen. The corpus contains the same docvars as the input. Email addresses or URLs are treated like nouns.
}
\description{
This function offers two algorithms for topic/content masking. See Details below.
}
\details{
References to algorithms to appear here.
}
\examples{
text <- "The elegant cat was forcefully put on the chair. cat@pets.com; http://quanteda.io/"
toy.corpus <- quanteda::corpus(text)
contentmask(toy.corpus, algorithm = "POSnoise", replace_non_ascii = FALSE)
}
