% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_corpus.R
\name{create_corpus}
\alias{create_corpus}
\title{Create a corpus}
\usage{
create_corpus(path, pos_tag = F)
}
\arguments{
\item{path}{A string containing the path to a folder of plain text files (ending in .txt) with their name structured as following: authorname_textname.txt (e.g. smith_text1.txt).}

\item{pos_tag}{A logical value (default is FALSE) determining whether the corpus should be Part of Speech tagged in preparation for using an algorithm for content masking.}
}
\value{
Either a \code{quanteda} corpus object with the authors' names as a docvar or a data frame containing the Part of Speech tagged corpus.
}
\description{
Function to read in data, turn it into a corpus, and potentially Part-of-Speech tag it.
}
\examples{
\dontrun{
path <- "path/to/data"
create_corpus(path)
create_corpus(path, pos_tag = T)
}
}
