% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_corpus.R
\name{create_corpus}
\alias{create_corpus}
\title{Create a corpus}
\usage{
create_corpus(path)
}
\arguments{
\item{path}{A string containing the path to a folder of plain text files (ending in .txt) with their name structured as following: authorname_textname.txt (e.g. smith_text1.txt).}
}
\value{
A \code{quanteda} corpus object with the authors' names as a docvar.
}
\description{
Function to read in data and turn it into a corpus object.
}
\examples{
\dontrun{
path <- "path/to/data"
create_corpus(path)
}
}
