% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impostors.R
\name{impostors}
\alias{impostors}
\title{Impostors Method}
\usage{
impostors(
  q.data,
  k.data,
  cand.imps,
  algorithm = "RBI",
  coefficient = "minmax",
  k = 300,
  m = 100,
  n = 25,
  features = F,
  cores = NULL
)
}
\arguments{
\item{q.data}{The questioned or disputed data, either as a corpus (the output of \code{\link[=create_corpus]{create_corpus()}}) or as a \code{quanteda} dfm (the output of \code{\link[=vectorize]{vectorize()}}).}

\item{k.data}{The known or undisputed data, either as a corpus (the output of \code{\link[=create_corpus]{create_corpus()}}) or as a \code{quanteda} dfm (the output of \code{\link[=vectorize]{vectorize()}}). More than one sample for a candidate author is accepted for all algorithms except IM.}

\item{cand.imps}{The impostors data for the candidate authors, either as a corpus (the output of \code{\link[=create_corpus]{create_corpus()}}) or as a \code{quanteda} dfm (the output of \code{\link[=vectorize]{vectorize()}}).}

\item{algorithm}{A string specifying which impostors algorithm to use, either "RBI", "KGI", or "IM".}

\item{coefficient}{A string indicating the coefficient to use, either "minmax" (default) or "cosine". This does not apply to the algorithm KGI, where the distance is "minmax".}

\item{k}{The \emph{k} parameters for the RBI algorithm. Not used by other algorithms. The default is 300.}

\item{m}{The \emph{m} parameter for the IM algorithm. Not used by other algorithms. The default is 100.}

\item{n}{The \emph{n} parameter for the IM algorithm. Not used by other algorithms. The default is 25.}

\item{features}{A logical value indicating whether the important features should be retrieved or not. The default is FALSE. This only applies to the RBI algorithm.}

\item{cores}{The number of cores to use for parallel processing (the default is one).}
}
\value{
The function will test all possible combinations of q texts and candidate authors and return a
data frame containing a score ranging from 0 to 1 representing the degree of confidence that the candidate is the author of the Q text. The data frame contains a column called "target" with a logical value which is TRUE if the author of the Q text is the candidate and FALSE otherwise. If the RBI algorithm is selected and the features parameter is TRUE then the data frame will also contain a column with the features that are likely to have had an impact on the score. The three data sets, \code{q.data}, \code{k.data}, and \code{cand.imps} must be disjunct in terms of the texts that they contain otherwise an error is returned.
}
\description{
This function runs the \emph{Impostors Method} for authorship verification.
}
\details{
More details here.
}
\examples{
q <- refcor.sample[1]
ks <- refcor.sample[2:3]
imps <- refcor.sample[4:9]
impostors(q, ks, imps, algorithm = "KGI")
}
