% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{performance}
\alias{performance}
\title{Performance evaluation}
\usage{
performance(training, test = NULL)
}
\arguments{
\item{training}{The data frame with the results to evaluate, typically the output of an authorship verification function, such as \code{\link[=impostors]{impostors()}}. If only training is present then the function will perform a leave-one-out cross-validation.}

\item{test}{Optional data frame of results. If present then a calibration model is extracted from training and its performance is evaluated on this data set.}
}
\value{
The function returns a data frame with performance statistics, including the C_llr. The binary classification statistics are all calculated considering a Log-Likelihood Ratio score of 0 as a threshold.
}
\description{
This function is used to the test the performance of a verification function, such as the \emph{Impostors Method}.
}
\details{
More details here.
}
\examples{
results <- data.frame(score = c(0.5, 0.2, 0.8, 0.01), target = c(TRUE, FALSE, TRUE, FALSE))
performance(results)
}
